BasicUI works!
I'm not sure why though... I made a separate ViewController to interact with the TextField and Button and made an instance of it in the view. My guess is the instance let the objects stay around more permanently than putting references in the AppDelegate, which in turn lets the button have a non-nil target (action?).